move-historic/requirements.txt
google-cloud-storage==2.10.0
google-cloud-secret-manager==1.0.2
requests==2.30.0